pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Alaa Fawzy <alaafawzy963@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django-rest-knox = "^4.2.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.2"}
django = "^4.1.5"
django-encrypted-field = "^1.0.5"
django-searchable-encrypted-fields = "^0.2.1"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
